feat(l2): implement l1 fee
