+++ crates/common/types/l2.rs
@@ -1,2 +1,3 @@
+pub mod account_diff;
 pub mod batch;
 pub mod fee_config;
