+++ crates/vm/levm/src/hooks/hook.rs
@@ -29,7 +29,10 @@ pub fn l1_hooks() -> Vec<Rc<RefCell<dyn Hook + 'static>>> {
 
 pub fn l2_hooks(fee_config: FeeConfig) -> Vec<Rc<RefCell<dyn Hook + 'static>>> {
     vec![
-        Rc::new(RefCell::new(L2Hook { fee_config })),
+        Rc::new(RefCell::new(L2Hook {
+            fee_config,
+            backup_hook: BackupHook::default(),
+        })),
         Rc::new(RefCell::new(BackupHook::default())),
     ]
 }
